requirements.txt
anyio==3.6.2
click==8.1.3
coloredlogs==15.0.1
fastapi==0.95.0
flatbuffers==23.3.3
h11==0.14.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
mpmath==1.3.0
numpy==1.24.2
onnxruntime==1.14.1
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
protobuf==4.22.1
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
